more changes

+++ src/main/java/guru/data/property/price/ingest/handler/model/property/GeoLocation.java
@@ -6,6 +6,7 @@ import lombok.Value;
 @Builder
 @Value
 public class GeoLocation {
+
   String type;
-  float[] coordinates;
+  double[] coordinates;
 }
